# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "modules Platform API",
    "version": "1.0.0",
    "description": "REST API for interacting the modules platform.",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/your-org/modules-repo"
    },
    "license": {
      "name": "Unlicense",
      "url": "https://unlicense.org/"
    }
  },
  "servers": [
    {
      "url": "{protocol}://{host}/api",
      "description": "Configurable server",
      "variables": {
        "protocol": {
          "enum": ["http", "https"],
          "default": "http"
        },
        "host": {
          "default": "localhost:3000",
          "description": "Server host and port"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session-based authentication using JWT tokens from Better Auth"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key authentication for programmatic access. Supports 'read' and 'write' scopes."
      }
    },
    "schemas": {
...
      "ModuleResponseSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique module ID"
          },
          "name": {
            "type": "string",
            "description": "Module name"
          },
          "shortDescription": {
            "type": "string",
            "description": "Brief module description"
          },
          "description": {
            "type": "string",
            "description": "Full module description"
          },
          "author": {
            "type": "string",
            "description": "Module author"
          },
          "category": {
            "type": "string",
            "description": "Module category"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          },
          "icon": {
            "type": "string",
            "nullable": true,
            "description": "Module icon URL"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "description": "Screenshot URLs"
          },
          "isOpenSource": {
            "type": "boolean",
            "description": "Open source status"
          },
          "license": {
            "type": "string",
            "description": "License type"
          },
          "compatibility": {
            "type": "object",
            "properties": {
              "androidVersions": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "rootMethods": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": ["androidVersions", "rootMethods"],
            "description": "Compatibility information"
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Feature list"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true,
            "description": "Source code repository URL"
          },
          "communityUrl": {
            "type": "string",
            "nullable": true,
            "description": "Community/support URL"
          },
          "isFeatured": {
            "type": "boolean",
            "description": "Featured status"
          },
          "isRecommended": {
            "type": "boolean",
            "description": "Recommended status"
          },
          "downloads": {
            "type": "number",
            "nullable": true,
            "description": "Total download count"
          },
          "rating": {
            "type": "number",
            "nullable": true,
            "description": "Average rating (1-5)"
          },
          "ratingCount": {
            "type": "number",
            "nullable": true,
            "description": "Number of ratings"
          }
        },
        "required": [
          "id",
          "name",
          "shortDescription",
          "description",
          "author",
          "category",
          "lastUpdated",
          "isOpenSource",
          "license",
          "compatibility",
          "features",
          "isFeatured",
          "isRecommended"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/trending": {
      "get": {
        "operationId": "get-trending",
        "summary": "Get trending modules",
        "description": "Discover trending modules using sophisticated algorithms based on recent downloads, ratings, and updates. Multiple algorithms available.",
        "tags": ["s Trending"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved trending modules with detailed trend analysis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "modules": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ModuleResponseSchema"
                      },
                      "description": "Trending modules"
                    }
                  },
                  "required": ["modules"]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}